+++ Rdmp.Core/DataFlowPipeline/PipelineCrashedException.cs
@@ -8,6 +8,9 @@ using System;
 
 namespace Rdmp.Core.DataFlowPipeline
 {
+    /// <summary>
+    /// Exception thrown when an <see cref="DataFlowPipelineEngine{T}"/> crashes with a fatal exception
+    /// </summary>
     public class PipelineCrashedException : Exception
     {
         public PipelineCrashedException(string msg, Exception innerException):base(msg,innerException)